webcam-server/requirements.txt
Events==0.3
imutils==0.4.6
Flask==0.12.2
opencv_python==3.4.0.12
paho_mqtt==1.3.1
